use common js modules for react components (it worksgit add -p)
