+++ app/assets/javascripts/application.js
@@ -17,7 +17,6 @@
 //= require react
 //= require react_ujs
 //= require fluxxor
-//= require_tree ./flux
 //= require components
 //= require bootstrap-sprockets
 //= require select2
