+++ app/assets/javascripts/components/widgets/text_body_widget.js.jsx
@@ -1,3 +1,6 @@
+var Widget = require('components/widgets/widget')
+var TextBodyWidgetForm = require('components/widgets/text_body_widget_form')
+
 var TextBodyWidget = React.createClass({
 
   propTypes: {
@@ -31,41 +34,6 @@ var TextBodyWidget = React.createClass({
       </div>
     )
   }
-})
+});
 
-var TextBodyWidgetForm = React.createClass({
-
-  propTypes: {
-    text_body_html:   React.PropTypes.string,
-    campaign_page_id: React.PropTypes.number.isRequired,
-    id:               React.PropTypes.number
-  },
-
-  mixins: [FluxMixin],
-
-  handleSubmit(e) {
-    e.preventDefault()
-    var text = React.findDOMNode(this.refs.body).value
-    var data = {text_body_html: text, text: text, type: 'TextBodyWidget', campaign_page_id: this.props.campaign_page_id }
-    if ('id' in this.props) {
-      data.id = this.props.id;
-      this.getFlux().actions.updateWidget(data);
-    } else {
-      this.getFlux().actions.createWidget(data);
-    }
-  },
-
-  render() {
-    return (
-      <div className='widget-html-form'>
-         <form onSubmit={ this.handleSubmit }>
-          <div className="form-group">
-            <label htmlFor="">Text</label>
-            <textarea className='form-control' ref='body' defaultValue={this.props.text_body_html}></textarea>
-          </div>
-          <button type="submit" className="btn btn-default">Submit</button>
-        </form>
-      </div>
-    )
-  }
-})
+module.exports = TextBodyWidget;
