+++ app/assets/javascripts/flux/mixins.js
@@ -1,0 +1,6 @@
+var mixins = {
+  FluxMixin: Fluxxor.FluxMixin(React),
+  StoreWatchMixin: Fluxxor.StoreWatchMixin
+}
+
+module.exports = mixins;
